Collect Log with java agent (#6127)
