+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-logback-1.x-activation/pom.xml
@@ -38,5 +38,10 @@
             <version>${logback-classic.version}</version>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>org.apache.skywalking</groupId>
+            <artifactId>apm-toolkit-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 </project>
