+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-logback-1.x-activation/src/main/java/org/apache/skywalking/apm/toolkit/activation/log/logback/v1/x/log/GRPCLogAppenderInterceptor.java
@@ -25,7 +25,6 @@ import java.util.Optional;
 import org.apache.skywalking.apm.agent.core.boot.ServiceManager;
 import org.apache.skywalking.apm.agent.core.conf.Config;
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
-import org.apache.skywalking.apm.agent.core.log.LogReportServiceClient;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
@@ -36,6 +35,8 @@ import org.apache.skywalking.apm.network.logging.v3.LogDataBody;
 import org.apache.skywalking.apm.network.logging.v3.TextLog;
 import org.apache.skywalking.apm.network.logging.v3.TraceContext;
 import org.apache.skywalking.apm.network.logging.v3.YAMLLog;
+import org.apache.skywalking.apm.toolkit.common.log.GRPCLogConfig;
+import org.apache.skywalking.apm.toolkit.common.log.LogReportServiceClient;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 
@@ -92,7 +93,7 @@ public class GRPCLogAppenderInterceptor implements InstanceMethodsAroundIntercep
 
         LogDataBody.Builder logDataBodyBuilder = LogDataBody.newBuilder();
         LogDataBody.ContentCase contentType = Optional.ofNullable(LogDataBody.ContentCase.forNumber(
-                Config.GRPCLog.LOG_BODY_TYPE)).orElse(LogDataBody.ContentCase.TEXT);
+                GRPCLogConfig.Plugin.GRPCLog.LOG_BODY_TYPE)).orElse(LogDataBody.ContentCase.TEXT);
         switch (contentType) {
             case JSON:
                 logDataBodyBuilder.setType(LogDataBody.ContentCase.JSON.name())
