+++ apm-sniffer/config/agent.config
@@ -112,8 +112,8 @@ plugin.mount=${SW_MOUNT_FOLDERS:plugins,activations}
 # plugin.springannotation.classname_match_regex=${SW_SPRINGANNOTATION_CLASSNAME_MATCH_REGEX:}
 
 #grpc-log client configuration
-grpclog.service_host=${SW_GRPC_LOG_SERVER_HOST:localhost}
-grpclog.service_port=${SW_GRPC_LOG_SERVER_PORT:8000}
-grpclog.max_message_size=${SW_GRPC_LOG_MAX_MESSAGE_SIZE:10485760}
-grpclog.log_body_type=${SW_GRPC_LOG_LOG_BODY_TYPE:2}
-grpclog.upstream_timeout=${SW_GRPC_LOG_GRPC_UPSTREAM_TIMEOUT:30}
+# plugin.grpclog.service_host=${SW_GRPC_LOG_SERVER_HOST:localhost}
+# plugin.grpclog.service_port=${SW_GRPC_LOG_SERVER_PORT:8000}
+# plugin.grpclog.max_message_size=${SW_GRPC_LOG_MAX_MESSAGE_SIZE:10485760}
+# plugin.grpclog.log_body_type=${SW_GRPC_LOG_LOG_BODY_TYPE:2}
+# plugin.grpclog.upstream_timeout=${SW_GRPC_LOG_GRPC_UPSTREAM_TIMEOUT:30}
